bathos_unix_defconfig.conf: define __printf on cmdline
printf() is a macro for bathos:

This is to cope with the atmega case, which would otherwise require every
printf fmt string to be written like this:

printf(PSTR("hello world\n"))

to have the string stored into .text instead of .data (and save a lot of RAM).

We use a linker flag to redirect __printf to printf, so that the original
bathos allocator object files can be linked to the test program object
file without getting any "undefined reference to __printf" error.